Add: Greedy loading of request body
By default, werkzeug uses immutable data types for storing request
values. This causes a problem for middlewares that need to modify
the request object.

In this case, the request body needed to be mutable in order to
provide io middleware. Now a mutable request body is stored in a
property named 'body' attached a Request object.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>